Refactor: 컨텐트 노말 컴포넌트 리팩토링
