+++ frontend/src/components/contents/NormalContent/style.ts
@@ -1,14 +1,12 @@
 import styled from '@emotion/styled';
 
-import { POST_CONTENT_WIDTH } from 'src/globals/constants';
-
 const Wrapper = styled.div`
   margin: 16px;
-  width: ${POST_CONTENT_WIDTH};
 `;
 
 const Content = styled.p`
   word-wrap: break-word;
+  white-space: pre-wrap;
 `;
 
 export { Wrapper, Content };
